fix: eliminate all clippy warnings, harden edge-device safety
Clippy (78 → 0 warnings):
- Replace manual div_ceil, is_multiple_of with stdlib methods
- Simplify match → if let, map_or, unwrap_or_default patterns
- Fix reference-to-reference patterns in i-Octree
- Use iterator patterns instead of index-based loops
- Add type aliases for complex types in FTS and UnifiedMemTable

Performance & safety fixes from code review:
- scan_range: propagate I/O errors instead of silently swallowing
  them via filter_map(|r| r.ok()) (P1 — hides SD card corruption)
- zstd decompression: increase max output from 1MB to 4MB (P1 —
  blocks with large inline values would fail silently on edge)
- row_id allocation: Relaxed → SeqCst ordering for ARM safety (P2)
- batch_insert: same SeqCst fix for bulk row_id allocation (P2)
- Remove dead identical-if-block in text search scoring

+++ src/storage/columnar/store.rs
@@ -684,7 +684,7 @@ impl ColumnarStore {
 
         for segment in &segments {
             let mut col_ids: Vec<u16> = column_ids.iter().map(|(id, _)| *id).collect();
-            let row_id_col = segment.column_count as u16;
+            let row_id_col = segment.column_count;
             if segment.has_row_id_column {
                 col_ids.push(row_id_col);
             }
@@ -692,7 +692,7 @@ impl ColumnarStore {
             let blocks = manager.read_columns(segment, &col_ids)?;
 
             let schema_blocks: Vec<&ColumnBlock> = blocks.iter()
-                .filter(|b| b.column_id < segment.column_count as u16)
+                .filter(|b| b.column_id < segment.column_count)
                 .collect();
             let decoded = self.decode_columns(&schema_blocks, &schema, segment.row_count as usize)?;
 
@@ -921,7 +921,7 @@ impl ColumnarStore {
         for segment in &segments {
             // P1: Read schema columns + row_id column (column_count = one past last schema col)
             let mut col_ids: Vec<u16> = column_ids.iter().map(|(id, _)| *id).collect();
-            let row_id_col = segment.column_count as u16; // row_ids stored as extra column
+            let row_id_col = segment.column_count; // row_ids stored as extra column
             if segment.has_row_id_column {
                 col_ids.push(row_id_col);
             }
@@ -964,7 +964,7 @@ impl ColumnarStore {
 
             // Now decode only needed columns for matching rows
             let schema_blocks: Vec<&ColumnBlock> = blocks.iter()
-                .filter(|b| b.column_id < segment.column_count as u16)
+                .filter(|b| b.column_id < segment.column_count)
                 .collect();
             let decoded = self.decode_columns(&schema_blocks, &schema, segment.row_count as usize)?;
